fix(tools): serialize runtime-proxy tests across the whole tools crate
The proxy-bypass execute() path now reads the process-global runtime proxy
config, but PROXY_TEST_LOCK was private to file_download::tests, so only that
module serialized on it. Sibling real-request upload/bundle tests that build
clients from the same global and proxy_config tests that write it could race
with a temporarily-installed proxy, making the full zeroclaw-tools lib suite
order-dependent (e.g. build_secure_download_client_ignores_runtime_http_proxy
saw 2 requests).

Promote the lock and RuntimeProxyGuard into a crate-wide cfg(test)
test_support module and acquire the shared lock in every test that reads,
writes, or builds a request from runtime proxy state.
